python3-scipy: Adapt the build if meta-scikit-learn is also used
With a newer python3-scipy version that relies on meson
to be built, and implicitly, on pkgconfig to detect the
presence of dependencies, this error occurs when both
meta-scipy and meta-scikit-learn is present:

| ../scipy-1.11.3/scipy/meson.build:161:9: ERROR: Dependency lookup for lapack with method 'pkgconfig' failed: Could not generate cargs for lapack:
| Package blas was not found in the pkg-config search path.
| Perhaps you should add the directory containing `blas.pc'
| to the PKG_CONFIG_PATH environment variable
| Package 'blas', required by 'lapack', not found

This is caused by the inconsistency of lapack_%.bbappend present in
both meta-scipy (which builds lapack with its own internal blas
library) and meta-scikit-learn (which in turn builds lapack with
openblas).

Adapt python3-scipy to use openblas in this case, so there is
no conflict.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>